Allow connecting to simulation devices by ssh through virl host.
Previously, even in case if there is a mgmt-lxc container in the
simulation lab, we had to connect to the mgmt-lxc directly, inspite of
we can connect to mgmt-lxc through the virl host.

In my opinion, as we need to be able to reach virl host anyway it is
better to connect to mgmt-lxc through it by default.